fix: ref warning from RefToTgphRef (#247)
### Description
When in dev mode in control, the console became littered with the warning `Function components cannot be given refs. Attempts to access this ref will fail. Did you mean to use React.forwardRef()`. This was happening because we were always passing the `ref` from `React.fowardRef` into the underlying component via radix `<Slot/>`. This PR refactors the `RefToTgphRef` component to detect if the children contain `React.forwardRef` and ONLY pass `ref` under that circumstance. This totally eliminates the warning while still maintaining the ability to pass `ref` to components outside of the telegraph ecosystem. Doing this refactor will also allow us to more granularly control how refs are passed when we eventually need to support React v19.

### Tasks
[KNO-6839](https://linear.app/knock/issue/KNO-6839/[telegraph]-fix-ref-warning-stemming-from-reftotgphref)