chore: upgrades Vite plugins, fixes peer dependency issues (#407)
# chore: upgrades Vite plugins, fixes peer dependency issues

In an effort to support React 19 with out JS SDK, we have to make sure Telegraph can support React 19 as well, since its a dependency there.

[Here is the major upgrade guide for `vanilla-extract`](https://github.com/vanilla-extract-css/vanilla-extract/releases/tag/%40vanilla-extract%2Fvite-plugin%405.0.0)

When running `yarn install` on main:

```bash
➤ YN0060: │ vite is listed by your project with version 6.0.11, which doesn't satisfy what @vanilla-extract/vite-plugin (p5b4dc) and other dependencies request (^4.2.0 || ^5.0.0).
```

A step in the process to address [#317](https://github.com/knocklabs/javascript/issues/317) and [KNO-7523](https://linear.app/knock/issue/KNO-7523/[js]-support-react-19-in-react-sdks).

As a first step, I'm going to update the existing peer dependency issues. Next will be react versions.

## Open question about workflow

I'm trying to use graphite for the first time (still learning...). I wanted to create a stack of PRs that merge into a primary branch (would be this one that leads into main). This work also isn't tracked in linear outside the root ticket for the JS SDK (should I be making a ticket for a PR like this?). So figuring this workflow out right now, forgive me if the branch naming is bumpy.