feat: update icon component lucide loading strategy to decrease bundle size (#498)
### Description
When exporting the `Lucide` object from `@telegraph/icon` it would include every icon in the bundle package bundle and the bundle of the applications using it. Even with the new approach which utilized `lucide-react/dynamic` each icon would end up in it's own chunk. This meant the bundle size of apps using `@telegraph/icon` were WAY larger than they should be. Now, we import icons directly from `lucide-react` so that a bundler can properly treeshake which icons are being used in a project and only include those icons in the final bundle. This will have breaking changes on any package that import the `Lucide` object, but because we are no longer exporting it the build will fail resulting in no risk in pushing out incorrect code. In a large project like our dashboard, we'll update this functionality with a codemod. 

- Removes `Lucide` export from `@telegraph/icon`.
- Updates `<Icon/>` to previous version where the lucide icon react component was used directly.
- Updates `Lucide` object usage across telegraph to instead import icons directly from `lucide-react`.
- Installs `lucide-react` as a dependency in packages that utilize it. 
- Bumps `@telegraph/icon` by a minor version and the other effect versions with a patch version.